Add a convenient command to start only modified services
